ci: enforce cargo fmt + clippy on push to main
New `.github/workflows/ci.yml` runs three jobs on every push to
main and every PR targeting main, all with `-D warnings`:

- `cargo fmt --all -- --check` (rustfmt parity)
- `cargo clippy --workspace --exclude pocopine-cli --target
  wasm32-unknown-unknown --all-targets` (wasm crates)
- `cargo clippy -p pocopine-cli --all-targets` (host CLI)

The clippy step doesn't allow warnings, so this commit also
clears every existing lint across the workspace so main lands
green:

- `pocopine-core` for_plan / for_: `as_ref().map(deref)` →
  `as_deref()`, `let-else None` → `?`, `iter().any` →
  `contains`, `let _ = unit` → bare statement, `+ patches DOM`
  doc rewrite (was being parsed as a markdown list item).
- `pocopine-macros`: `eq_ignore_ascii_case` for tag matching,
  `find_attr(...)?`, `&[u16]` instead of `&Vec<u16>`,
  `#[allow(clippy::only_used_in_recursion)]` on the AST walkers.
- `pine-motion` Easing + `pine` WeekDayFormat → derive
  `Default` with `#[default]`.
- `pine-icons` `&'static str` → `&str` in const tuple type.
- `pine` datetime comparators: `map_or(false, ...)` →
  `is_some_and(...)`.
- `pine` `TimeSegments::from_str` renamed to `parse` (avoid
  confusion with `std::str::FromStr::from_str`).
- `pine` date_field: drop needless borrow on `emit_model_field`.
- `website` calendar demo: drop unnecessary `as u32` casts.
- `pocopine-macros` `template_parser`: elide the explicit `'a`
  on a test helper.

Plus rustfmt-canonical reformats across ~20 unrelated files
that had drifted (single-line vs multi-line `use` statements) —
folded into this commit to keep the workspace fmt-clean under
the new gate.

+++ examples/website/src/components/showcase/calendar/mod.rs
@@ -70,8 +70,8 @@ impl CalendarDemo {
         {
             let js = js_sys::Date::new_0();
             let y = js.get_full_year() as i32;
-            let m = (js.get_month() + 1) as u32;
-            let d = js.get_date() as u32;
+            let m = js.get_month() + 1;
+            let d = js.get_date();
             self.placeholder = DateValue::new(y, m as u8, d as u8);
         }
     }
